==> workflow/00-import.Rmd <==
---
title: "Import Data"
date: "Compiled at `r format(Sys.time(), '%Y-%m-%d %H:%M:%S', tz = 'UTC')` UTC"
output: github_document
params:
  name: "00-import" # change if you rename file
---

```{r here, message=FALSE}
here::i_am(paste0(params$name, ".Rmd"), uuid = "f8c9b430-542e-4eaa-b315-bad86866aa06")
```

The purpose of this document is to import the data we'll need to make some COVID-19 maps for the US:

- [NYT daily state-level data](https://github.com/nytimes/covid-19-data/blob/master/us-states.csv)
- [US state population estimates](https://github.com/JoshData/historical-state-population-csv/blob/primary/historical_state_population_by_year.csv) from [Josh Tauberer](https://github.com/JoshData).

```{r packages}
library("conflicted")
```

```{r directories}
# create or *empty* the target directory, used to write this file's data: 
projthis::proj_create_dir_target(params$name, clean = TRUE)

# function to get path to target directory: path_target("sample.csv")
path_target <- projthis::proj_path_target(params$name)

# function to get path to previous data: path_source("00-import", "sample.csv")
path_source <- projthis::proj_path_source(params$name)
```

## Download

We call `download.file()` to put the files directly into our target directory, using `path_target()` to specify the path within the target directory.
First, the COVID data:

```{r covid-states}
download.file(
  "https://raw.githubusercontent.com/nytimes/covid-19-data/master/us-states.csv",
  destfile = path_target("covid-states.csv")
)
```

Next, the US states population data:

```{r population-states}
download.file(
  "https://raw.githubusercontent.com/JoshData/historical-state-population-csv/primary/historical_state_population_by_year.csv",
  destfile = path_target("population-states.csv")
)
```

## Files written

These files have been written to ```r paste0("data/", params$name)```:

```{r list-files-target}
projthis::proj_dir_info(path_target())
```
