==> workflow/01-clean.Rmd <==
---
title: "Clean data"
date: "Compiled at `r format(Sys.time(), '%Y-%m-%d %H:%M:%S', tz = 'UTC')` UTC"
output: github_document
params:
  name: "01-clean" # change if you rename file
---

```{r here, message=FALSE}
here::i_am(paste0(params$name, ".Rmd"), uuid = "9fa9049e-5898-494b-9b1a-0175496b3975")
```

The purpose of this document is to clean the data we imported in the [previous step](00-import.md).

```{r packages}
library("conflicted")
library("readr")
library("dplyr")

conflict_prefer("filter", "dplyr")
```

```{r directories}
# create or *empty* the target directory, used to write this file's data: 
projthis::proj_create_dir_target(params$name, clean = TRUE)

# function to get path to target directory: path_target("sample.csv")
path_target <- projthis::proj_path_target(params$name)

# function to get path to previous data: path_source("00-import", "sample.csv")
path_source <- projthis::proj_path_source(params$name)
```

## Read data

First, we will read in the data:

```{r population_raw}
population_raw <- 
  read_csv(
    path_source("00-import", "population-states.csv"), 
    col_names = c("abbreviation", "year", "population"),
    col_types = cols(abbreviation = "c", year = "d", population = "d")
  ) %>%
  print()
```

```{r covid_raw}
covid_raw <- 
  read_csv(
    path_source("00-import", "covid-states.csv"),
    col_types = cols(
      date = col_date(format = ""),
      state = col_character(),
      fips = col_character(),
      cases = col_double(),
      deaths = col_double()
    )
  ) %>%
  print()
```

Reading in the data using `readr::read_csv()`, we use the `cols()` function with the `col_types` argument to assert the types of the columns that we parse.
If there is some data that is *not* of this sort, we want to know about it.

## Wrangle data

It looks like we could use a way to link state names with their abbreviations:

```{r states}
states <- 
  tibble(
    state = c(state.name, "District of Columbia"),
    abbreviation = c(state.abb, "DC")
  ) %>%
  print()
```
We want to work with:

- 2019 population
- 50 US states, plus District of Columbia

```{r population}
population <- 
  population_raw %>%
  filter(year == 2019) %>%
  right_join(states, by = "abbreviation") %>%
  select(state, population) %>%
  print()
```
```{r covid}
covid <- 
  covid_raw %>%
  filter(state %in% states$state) %>%
  print()
```

We can see which states have the most cases, also verifying the recency of the data:

```{r recent}
covid %>%
  filter(date == max(date)) %>%
  arrange(desc(cases))
```

## Write data

Let's write out our population and covid data:

```{r write-population}
write_csv(population, path_target("population.csv"))
```

```{r write-covid}
write_csv(covid, path_target("covid.csv"))
```

## Files written

These files have been written to ```r paste0("data/", params$name)```:

```{r list-files-target}
projthis::proj_dir_info(path_target())
```
